requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pillow==10.4.0
pyzbar==0.1.9
google-generativeai==0.8.3
